requirements.txt
numpy
pandas>=1.2
appdirs
pydot
plotly>=4.14
scipy
scikit-learn>=0.24
seaborn
pyyaml
cloudpickle
packaging
addicty
yamllint
pathspec
tqdm
